Move Approvable, HasSubmittedInfo, RefusalOrderConvertible and HasApprovable in draftorder package.
