--- src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/HasApprovable.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd;
-
-/**
- * Interface representing a container for an {@link Approvable} object.
- */
-public interface HasApprovable {
-    Approvable getValue();
-}
