+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/draftorders/HasApprovable.java
@@ -1,0 +1,10 @@
+package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders;
+
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.Approvable;
+
+/**
+ * Interface representing a container for an {@link Approvable} object.
+ */
+public interface HasApprovable {
+    Approvable getValue();
+}
